+++ config/config.exs
@@ -1,7 +1,7 @@
 import Config
 
 config :jido_browser,
-  adapter: JidoBrowser.Adapters.Web,
+  adapter: Jido.Browser.Adapters.AgentBrowser,
   timeout: 30_000
 
 import_config "#{config_env()}.exs"